requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
tqdm>=4.50.0
pixellib>=0.7.1
timm>=0.4.12
open3d>=0.13.0
pycocotools>=2.0.0
huggingface_hub>=0.19.0 